% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetPopulations.R
\name{compareNetworkPopulations}
\alias{compareNetworkPopulations}
\title{Evaluates gene differential regulation between population of networks based on mahalinobis distances.}
\usage{
compareNetworkPopulations(
  grnList,
  groupID,
  geneSets,
  scoringType = "gSet",
  test = wilcox.test()
)
}
\arguments{
\item{grnList}{A list of gene regulatory networks with transcription factors in the rows and target genes in the columns}

\item{groupID}{A vector with the group ID for each gene regulatory network}

\item{geneSets}{Gene sets provided as a \emph{list}}

\item{scoringType}{Either \emph{'gSet'} or \emph{'gene'}. If \emph{'gSet'} the single-sample Gene Set Enrichment Score (ssGSEA) for each gene set is returned. If \emph{'gene'} the mahalinobis distance is returned for each gene.}

\item{test}{The statistical test to be performed. By default uses the Wilcoxon Sum Rank Test, if more than two groups are about to be compared the Kruskal-Wallis Rank Sum Test should be used.}
}
\value{
A \emph{data.frame} single-sample gene set enrichment score for each gene set or the mahalinobis distance for each target gene in each network. P-values and FDR values are also provided.
}
\description{
Perform gene set scoring using a gene regulatory network as an input.
}
\author{
Daniel Osorio <dcosorioh@utexas.edu>
}
