% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{compareNetworks}
\alias{compareNetworks}
\title{Calculate differences between two gene regulatory networks}
\usage{
compareNetworks(X, Y)
}
\arguments{
\item{X}{GRN1}

\item{Y}{GRN2}
}
\description{
Using the mahalinobis distance, calculates the gene in-degree for two different PANDA gene regulatory networks.
}
\author{
Daniel Osorio <daniecos@uio.no>
}
