% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosEdges.R
\name{.drawCircosLegends}
\alias{.drawCircosLegends}
\title{Draw effect-size, novelty and gene-set legends on a Circos plot.}
\usage{
.drawCircosLegends(
  col_fun,
  valueRange,
  colorBy,
  hasPrior,
  knownColor,
  novelColor,
  set_colors,
  lwdRange = NULL,
  widthShown = FALSE,
  degCol_fun = NULL,
  degRange = NULL,
  degColBy = "log2FoldChange"
)
}
\description{
Draw effect-size, novelty and gene-set legends on a Circos plot.
}
\keyword{internal}
