% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosEdges.R
\name{.fetchGeneCoords}
\alias{.fetchGeneCoords}
\title{Download gene coordinates from Ensembl via biomaRt.}
\usage{
.fetchGeneCoords(genes, species, mirror)
}
\description{
Download gene coordinates from Ensembl via biomaRt.
}
\keyword{internal}
