% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosEdges.R
\name{.orderChr}
\alias{.orderChr}
\title{Natural-sort chromosome names (1..N, then X, Y, MT, then the rest).}
\usage{
.orderChr(chr)
}
\description{
Natural-sort chromosome names (1..N, then X, Y, MT, then the rest).
}
\keyword{internal}
