% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosEdges.R
\name{.validateGeneCoords}
\alias{.validateGeneCoords}
\title{Validate a user-supplied gene coordinate table.}
\usage{
.validateGeneCoords(geneCoords)
}
\description{
Validate a user-supplied gene coordinate table.
}
\keyword{internal}
