% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netScoring.R
\name{netScoring}
\alias{netScoring}
\title{Gene Set Network Scoring}
\usage{
netScoring(regNetwork, geneSets, scoreType = "gSet")
}
\arguments{
\item{regNetwork}{A gene regulatory network with transcription factors in the rows and target genes in the columns}

\item{geneSets}{Gene sets provided as a \emph{list}}

\item{scoreType}{Either \emph{'gSet'} or \emph{'gene'}. If \emph{'gSet'} the single-sample Gene Set Enrichment Score (ssGSEA) for each gene set is returned. If \emph{'gene'} the mahalinobis distance is returned for each gene.}
}
\value{
A named vector with the single-sample gene set enrichment score for each gene set or the mahalinobis distance for each target gene in the network.
}
\description{
Perform gene set scoring using a gene regulatory network as an input.
}
\author{
Daniel Osorio <dcosorioh@utexas.edu>
}
