% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpion.R
\name{scorpion}
\alias{scorpion}
\title{Constructs PANDA gene regulatory networks from single-cell gene expression data}
\usage{
scorpion(motif = NULL, expr = NULL, ppi = NULL)
}
\arguments{
\item{motif}{TF Motif}

\item{expr}{Single-cell Expression Data}

\item{ppi}{PPI}
}
\value{
A list
}
\description{
Constructs gene regulatory networks from single-cell gene expression data using the PANDA (Passing Attributes between Networks for Data Assimilation) algorithm
}
\author{
Daniel Osorio <daniecos@uio.no>
}
